pyproject.toml
[tool.poetry]
name = "reid"
version = "0.1.0"
description = ""
authors = ["grungy <20245612+grungy@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
torch = ">=2.6.0,<3.0.0"
torchvision = ">=0.21.0,<0.22.0"
matplotlib = ">=3.10.3,<4.0.0"
numpy = ">=1.24.0,<2.0.0"
requests = ">=2.32.4,<3.0.0"
pillow = ">=11.3.0,<12.0.0"
transformers = ">=4.46.0,<5.0.0"
pathlib = ">=1.0.1,<2.0.0"
llama-index = "^0.12.28"
llama-index-vector-stores-milvus = "^0.7.3"
pymilvus = ">=2.4.0,<3.0.0"
mlflow = "^3.1.4"
boto3 = "^1.40.1"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
